[Add] Confirmation dialog for delete

+++ lib/src/pages/administration/admin_challenges_page/widgets/admin_challenge_card.dart
@@ -1,6 +1,7 @@
 import 'package:appli_wei_custom/models/administration/admin_challenge.dart';
 import 'package:appli_wei_custom/src/pages/administration/admin_challenge_edit_page/admin_challenge_edit_page.dart';
 import 'package:appli_wei_custom/src/providers/admin_challenges_store.dart';
+import 'package:appli_wei_custom/src/shared/dialogs/delete_confirmation_dialog.dart';
 import 'package:appli_wei_custom/src/shared/widgets/challenge_images/admin_challenge_image.dart';
 import 'package:appli_wei_custom/src/shared/widgets/wei_card.dart';
 import 'package:flutter/material.dart';
@@ -59,9 +60,16 @@ class AdminChallengeCard extends StatelessWidget {
                     child: InkWell(
                       splashColor: Colors.red,
                       onTap: () async {
-                        final AdminChallengesStore challengesStore = Provider.of<AdminChallengesStore>(context, listen: false);
+                        final bool delete = await showDialog<bool>(
+                          context: context,
+                          builder: (context) => DeleteConfirmationDialog()
+                        );
 
-                        challengesStore.deleteChallenge(challenge);
+                        if (delete != null && delete) {
+                          final AdminChallengesStore challengesStore = Provider.of<AdminChallengesStore>(context, listen: false);
+
+                          challengesStore.deleteChallenge(challenge);
+                        }
                       },
                       child: const SizedBox(width: 32, height: 32, child: Icon(Icons.delete, color: Colors.white,)),
                     ),
