+++ lib/src/pages/administration/admin_teams_page/widgets/admin_team_card.dart
@@ -1,6 +1,7 @@
 import 'package:appli_wei_custom/models/team.dart';
 import 'package:appli_wei_custom/src/pages/administration/admin_team_edit_page/admin_team_edit_page.dart';
 import 'package:appli_wei_custom/src/providers/admin_teams_store.dart';
+import 'package:appli_wei_custom/src/shared/dialogs/delete_confirmation_dialog.dart';
 import 'package:appli_wei_custom/src/shared/widgets/team_image.dart';
 import 'package:appli_wei_custom/src/shared/widgets/wei_card.dart';
 import 'package:flutter/material.dart';
@@ -46,9 +47,16 @@ class AdminTeamCard extends StatelessWidget {
                 child: InkWell(
                   splashColor: Colors.red,
                   onTap: () async { 
-                    final AdminTeamsStore adminTeamsStore = Provider.of<AdminTeamsStore>(context, listen: false);
+                    final bool delete = await showDialog<bool>(
+                        context: context,
+                        builder: (context) => DeleteConfirmationDialog()
+                      );
 
-                    adminTeamsStore.deleteTeam(team);
+                    if (delete != null && delete) {
+                      final AdminTeamsStore adminTeamsStore = Provider.of<AdminTeamsStore>(context, listen: false);
+
+                      adminTeamsStore.deleteTeam(team);
+                    }
                   },
                   child: const SizedBox(width: 32, height: 32, child: Icon(Icons.delete, color: Colors.white,)),
                 ),
